more polish
