+++ apps/web/src/components/Layout.tsx
@@ -13,7 +13,7 @@ import {
 
 import { NavBar } from './NavBar';
 import { TopBar } from './TopBar';
-import { CreateFormTemplate } from './CreateFormTemplate';
+import { CreateFormTemplateModal } from './createFormTemplate/CreateFormTemplateModal';
 
 // Common layout component for all pages
 export const Layout = ({ children }: { children: any }) => {
@@ -40,26 +40,10 @@ export const Layout = ({ children }: { children: any }) => {
           {children}
         </Box>
       </Box>
-      <Modal
-        isOpen={isCreateFormTemplateOpen}
-        onClose={onCloseCreateFormTemplate}
-      >
-        <ModalOverlay />
-        <ModalContent minWidth="fit-content" height="fit-content">
-          <ModalCloseButton />
-          <ModalBody>
-            <CreateFormTemplate />
-          </ModalBody>
-          <ModalFooter>
-            <Button variant="ghost" mr={3} onClick={onCloseCreateFormTemplate}>
-              Cancel
-            </Button>
-            <Button color="#4C658A" onClick={() => {}}>
-              Create
-            </Button>
-          </ModalFooter>
-        </ModalContent>
-      </Modal>
+      <CreateFormTemplateModal
+        isCreateFormTemplateOpen={isCreateFormTemplateOpen}
+        onCloseCreateFormTemplate={onCloseCreateFormTemplate}
+      />
     </Box>
   );
 };
