+++ apps/web/src/components/createFormTemplate/types.tsx
@@ -1,0 +1,4 @@
+export type TempSignatureField = {
+  id: string;
+  value: string;
+};
