+++ apps/web/src/static/icons.tsx
@@ -164,3 +164,15 @@ export const EditIcon = createIcon({
   viewBox: '0 0 16 16',
   d: 'M0.368591 12.2309V15.3555H3.49316L12.7085 6.14008L9.58398 3.01552L0.368591 12.2309ZM15.1249 3.72375C15.4498 3.3988 15.4498 2.87387 15.1249 2.54891L13.1751 0.599185C12.8502 0.27423 12.3253 0.27423 12.0003 0.599185L10.4755 2.12397L13.6001 5.24854L15.1249 3.72375Z',
 });
+
+export const AddIcon = createIcon({
+  displayName: 'AddIcon',
+  viewBox: '0 0 19 19',
+  path: (
+    <>
+      <circle cx="9.5" cy="9.5" r="9.5" fill="#4C658A" />
+      <line x1="5" y1="9.5" x2="14" y2="9.5" stroke="white" />
+      <line x1="9.5" y1="5" x2="9.5" y2="14" stroke="white" />
+    </>
+  ),
+});
